Update -ForEach in tests

+++ scripts/uninstall/Uninstall-Greenshot.ps1
@@ -12,19 +12,21 @@ catch {
     Write-Warning -Message "Failed to stop Greenshot processes."
 }
 
-$params = @{
-    FilePath     = "$env:ProgramFiles\Greenshot\unins000.exe"
-    ArgumentList = "/VERYSILENT"
-    NoNewWindow  = $True
-    PassThru     = $True
-    Wait         = $True
-    ErrorAction  = "Continue"
-}
-$result = Start-Process @params
-$result.ExitCode
-if ($result.ExitCode -eq 0) {
-    if (Test-Path -Path "$env:ProgramFiles\Greenshot") {
-        Remove-Item -Path "$env:ProgramFiles\Greenshot" -Recurse -Force -ErrorAction "SilentlyContinue"
+if (Test-Path -Path "$env:ProgramFiles\Greenshot\unins000.exe") {
+    $params = @{
+        FilePath     = "$env:ProgramFiles\Greenshot\unins000.exe"
+        ArgumentList = "/VERYSILENT"
+        NoNewWindow  = $True
+        PassThru     = $True
+        Wait         = $True
+        ErrorAction  = "Continue"
+    }
+    $result = Start-Process @params
+    $result.ExitCode
+    if ($result.ExitCode -eq 0) {
+        if (Test-Path -Path "$env:ProgramFiles\Greenshot") {
+            Remove-Item -Path "$env:ProgramFiles\Greenshot" -Recurse -Force -ErrorAction "SilentlyContinue"
+        }
     }
 }
 #endregion
